Move converse-dragresize plugin to folder
