+++ src/plugins/dragresize/templates/dragresize.html
